+++ src/html_vars.js
